# docs/swagger.yaml
basePath: /
definitions:
  lib.Response:
    properties:
      message:
        type: string
      result: {}
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for User Management
  title: Backend CRUD API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Login using email and password
      parameters:
      - description: User Email
        format: email
        in: formData
        name: email
        required: true
        type: string
      - description: User Password
        format: password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Login
      tags:
      - Auth
  /register:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Create a new user account
      parameters:
      - description: User name
        in: formData
        name: name
        required: true
        type: string
      - description: User Email
        format: email
        in: formData
        name: email
        required: true
        type: string
      - description: User Password
        format: password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Register a new User
      tags:
      - Auth
  /upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload file Image
      parameters:
      - description: Image File
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - BearerAuth: []
      summary: Upload File
      tags:
      - Upload
  /users:
    get:
      description: Retrieve all User
      parameters:
      - default: 1
        description: Page Number
        in: query
        name: page
        type: integer
      - default: 5
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by name
        in: query
        name: search[name]
        type: string
      - description: Search by email
        in: query
        name: search[email]
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - BearerAuth: []
      summary: Get all User
      tags:
      - users
    post:
      consumes:
      - multipart/form-data
      description: Create New User with profile picture
      parameters:
      - description: User Name
        in: formData
        name: name
        required: true
        type: string
      - description: User email
        format: email
        in: formData
        name: email
        required: true
        type: string
      - description: User password
        format: password
        in: formData
        name: password
        required: true
        type: string
      - description: Profile Picture
        in: formData
        name: picture
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - BearerAuth: []
      summary: Create New User
      tags:
      - users
  /users/{id}:
    delete:
      description: Delete user by ID
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - BearerAuth: []
      summary: Delete User
      tags:
      - users
    get:
      description: Retrieve user Id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - BearerAuth: []
      summary: Get User by Id
      tags:
      - users
    put:
      consumes:
      - multipart/form-data
      description: Update User
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User nama
        in: formData
        name: name
        type: string
      - description: User email
        format: email
        in: formData
        name: email
        type: string
      - description: User password
        format: password
        in: formData
        name: password
        type: string
      - description: profile picture
        in: formData
        name: picture
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"